Agregue sweetalert

+++ src/app/admin/components/products-list/products-list.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import {ProductsService} from './../../../core/services/products/products.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-products-list',
@@ -25,10 +26,28 @@ export class ProductsListComponent implements OnInit {
   }
 
   deleteProduct(id: string){
-    this.productsService.deleteProduct(id)
-    .subscribe(rta=>{
-      this.fetchProducts();
-    });
+    Swal.fire({
+      title: '¿Estas seguro?',
+      text: "¿Deseas eliminar este elemento de forma permanente?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.productsService.deleteProduct(id)
+        .subscribe(rta=>{
+          this.fetchProducts();
+        });
+        Swal.fire(
+          'Eliminado!',
+          'El producto ha sido eliminado',
+          'success'
+        )
+      }
+    })
+  
   }
 
 }
